# src/Endpoint/VMware.ScriptRuntimeService.APIGateway/swagger.yaml
openapi: 3.0.1
info:
  title: Script Runtime Service for vSphere
  description: "# Script Runtime Service API\r\n\r\nScript Runtime Service for vSphere (SRS) allows running PowerShell and PowerCLI scripts. SRS is a VC add-on that is deployed separately from VCSA. SRS can be accessed via REST API that allows you to create PowerShell instances and run PowerShell and PowerCLI scripts within. No Connect-VIServer is required to run PowerCLI against VC(s) SRS is registered to.\r\n\r\n## Authetication\r\n\r\nSRS uses VC SSO as Identity and Authentication Server. Two types of authentication are supported. SIGN and Basic. SIGN authentication is purposed for Service-To-Service access to SRS resources. For the convenience of the end-users, SRS supports basic authentication passing username and password which are used to acquire SAML HoK token for SRS solution. When basic is used SRS exchanges the username and password for SAML HoK token from the SSO server. SRS uses the SAML token to Connect PowerCLI to VC services in further operations. \r\nOn successful authentication SRS returns API Key which is required to authorize further SRS API calls."
  version: '1.0'
  x-vmw-vapi-codegenconfig:
    package-name: com.vmware.srs
paths:
...
  /api/argument-scripts/templates:
    get:
      tags:
        - argumentscripts
      summary: List available argument script templates
      description: "### List available argument script templates\r\nArgument script templates are scripts with placeholders. When values replace the placeholders, the script can run in a given script runtime.\r\nArgument script templates help to convert simple type values to objects of types that can only be produced in a given script runtime. Those objects can be used as arguments to scripts' parameters.\r\n\r\nThis operation retrieves the available argument script templates."
      operationId: list.argument.scripts.templates
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArgumentScriptTemplate'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: argument_scripts
      x-vmw-vapi-methodname: list_templates
      x-vmw-doc-operation: list.argument.scripts.templates
    x-vmw-doc-resource: ArgumentScript
    x-vmw-doc-domain: argumentscripts
  '/api/argument-scripts/templates/{id}':
    get:
      tags:
        - argumentscripts
      summary: Retrieves argument script template by given unique template identifier
      description: "### Retrieves argument script template by given unique template identifier\r\nThis operation returns argument script template for the specified template id."
      operationId: get.argument.scripts.template
      parameters:
        - name: id
          in: path
          description: The Id of the argument script template
          required: true
          schema:
            type: string
            description: The Id of the argument script template
            nullable: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArgumentScriptTemplate'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: argument_scripts
      x-vmw-vapi-methodname: get_template
      x-vmw-doc-operation: get.argument.scripts.template
    x-vmw-doc-resource: ArgumentScript
    x-vmw-doc-domain: argumentscripts
  /api/argument-scripts/script:
    post:
      tags:
        - argumentscripts
      summary: Creates scripts for a given script template id and placeholder values
      description: "### Creates scripts for a given script template id and placeholder values\r\nReplaces the placeholders in a given argument transformation script template with given values on the placeholder_value_list field\r\nThe resulting script can be provided to a **script execution** parameter that expects a specific script runtime type\r\n\r\n### Example\r\nIf the template argument transformation script is\r\n\r\nGet-VM -Id <vm-id> -Server <server>\r\n\r\nThe result of this operation with given Id 'vm-1' and Server 'server-1' is\r\n\r\nGet-VM -Id 'vm-1' -Server 'server-1'"
      operationId: create.argument.scripts.script
      requestBody:
        description: The argument script create request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArgumentScript'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArgumentScript'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: argument_scripts
      x-vmw-vapi-methodname: create_script
      x-vmw-doc-operation: create.argument.scripts.script
    x-vmw-doc-resource: ArgumentScript
    x-vmw-doc-domain: argumentscripts
...
  /api/script-executions:
    post:
      tags:
        - scriptexecutions
      summary: Creates a script execution
      description: "### Create a script execution\r\n**Script execution** represents asynchronous execution of a script in a specified **runspace**\r\nWhen created **script execution** starts running in the **runspace**. To monitor the script execution\r\nprogress polling of the resource by its identifier should be used.\r\n### Retruns\r\nWhen the request is accepted **202 Accepted** with **Location** header is returned in the response that leads you to the **script execution** resource that is in running state initially.\r\nWhen script execution is requested with non-existing runspace  **404 Not Found** is returned."
      operationId: create.script.execution
      requestBody:
        description: Desired script execution resource.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptExecution'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptExecution'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: create
      x-vmw-doc-operation: create.script.execution
    get:
      tags:
        - scriptexecutions
      summary: List all script executions
      description: "### List all script executions      \r\n### Returns\r\nReturns a list of your script executions."
      operationId: list.script.executions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScriptExecution'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: list
      x-vmw-doc-operation: list.script.executions
    x-vmw-doc-resource: ScriptExecution
    x-vmw-doc-domain: scriptexecutions
  '/api/script-executions/{id}/cancel':
    post:
      tags:
        - scriptexecutions
      summary: Cancels a script execution
      description: "### Cancel a script execution\r\nThis operation is equivalent to pressing Ctrl+C in the PowerShell console. If the script is cancellable it will be canceled.\r\nThe state of the **script execution** will become canceled after this operation. The operation is asynchronous. Cancel request\r\nis sent to the runtime.\r\n\r\n### Returns\r\nThe operation doesn't return value. **200 Ok** will be returned if the request is successful."
      operationId: cancel.script.execution
      parameters:
        - name: id
          in: path
          description: The id of the script execution
          required: true
          schema:
            type: string
            description: The id of the script execution
            nullable: true
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: cancel
      x-vmw-doc-operation: cancel.script.execution
    x-vmw-doc-resource: ScriptExecution
    x-vmw-doc-domain: scriptexecutions
  '/api/script-executions/{id}':
    get:
      tags:
        - scriptexecutions
      summary: Retrieve a script execution
      description: "### Retrieve a script execution\r\nRetrieves the details of a **script execution**. You need only supply the unique script execution identifier that was returned on script execution creation.      \r\n### Returns\r\nReturns a **script execution** resource instance if a valid identifier was provided.\r\nWhen requesting the Id of a script execution that doesn't exist **404 NotFound** is returned."
      operationId: get.script.execution
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            nullable: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptExecution'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: get
      x-vmw-doc-operation: get.script.execution
    x-vmw-doc-resource: ScriptExecution
    x-vmw-doc-domain: scriptexecutions
components:
  schemas:
...
    ArgumentScriptTemplate:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the object.
          nullable: true
          readOnly: true
        script_runtime:
          type: string
          description: ScriptRuntime on which this script can be executed.
          nullable: true
          readOnly: true
        result_type:
          type: string
          description: Type name of the object that is produced by the script template.
          nullable: true
          readOnly: true
        script_template:
          type: string
          description: The script template.
          nullable: true
          readOnly: true
        placeholders:
          type: array
          items:
            type: string
          description: The script template placeholders that has to be replaced by strings to produce a valid script.
          nullable: true
      additionalProperties: false
      description: "ArgumentScriptTemplate object represents a script with placeholders. If placeholders are replaced by values, the script can be executed.\r\nThe purpose of the script is to produce an object of a type valid for a given script runtime. The result type of the script is usually\r\ninput type for other scripts. The ArgumentScriptTemplate scripts help to convert simple type values to \r\nobjects of types that can only be produced in a given script runtime."
    PlaceholderValue:
      required:
        - placeholder_name
        - value
      type: object
      properties:
        placeholder_name:
          type: string
          description: The name of the placeholder in the script template which will be replaced on the creation of an argument script.
        value:
          type: array
          items:
            type: string
          description: Value that will be used to replace the placeholder on the creation of an argument script.
      additionalProperties: false
      description: PlaceholderValue object represents a single value for a placeholder in template script.
    PlaceholderValueList:
      required:
        - values
      type: object
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/PlaceholderValue'
          description: List of script placeholder values.
      additionalProperties: false
      description: PlaceholderValueList object represents list of argument script placeholder values.
    ArgumentScript:
      required:
        - template_id
        - template_placeholder_value_list
      type: object
      properties:
        template_id:
          type: string
          description: Unique identifier for the argument template script.
        template_placeholder_value_list:
          type: array
          items:
            $ref: '#/components/schemas/PlaceholderValueList'
          description: "Placeholder value list which are used to create script from script template.\r\n\r\nSingle template_placeholder_value_list produces script by the given template replacing placeholder with the given values.\r\nMultiple items for template_placeholder_value_list produce a script of scripts which can produce an array of objects. Each template_placeholder_value_list item is used to produce script from template. Scripts are then combined in a multi-line script where each line produces result object."
        script:
          type: string
          description: Script result produced by the service based on given template_id and template_placeholder_parameters
          nullable: true
          readOnly: true
      additionalProperties: false
      description: ArgumentScript object represents a script produced from a given script template and placeholder values.
    ErrorDetails:
      type: object
      properties:
        code:
          type: integer
          description: Code of the internal service error. It is useful when error is reported to service maintainers.
          format: int32
        error_message:
          type: string
          description: Error message.
          nullable: true
        details:
          type: string
          description: Details of the error.
          nullable: true
      additionalProperties: false
      description: ErrorDetails object gives information for an API error.
      readOnly: true
...
    OutputObjectsFormat:
      enum:
        - text
        - json
      type: string
    ScriptExecutionState:
      enum:
        - success
        - error
        - running
        - canceled
      type: string
      description: ScriptExecutionState defines possible script execution states.
      readOnly: true
    ScriptExecution:
      required:
        - runspace_id
        - script
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the object.
          nullable: true
          readOnly: true
        runspace_id:
          type: string
          description: Unique identifier of the runspace where script execution is performed.
        name:
          type: string
          description: 'Name of the script execution. It is optional to give a name of the script execution on create request. If name was not specified on the script execution creation, the field has null value.'
          nullable: true
        script:
          type: string
          description: Content of the script.
        script_parameters:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/BooleanArrayScriptParameter'
              - $ref: '#/components/schemas/BooleanScriptParameter'
              - $ref: '#/components/schemas/IntegerArrayScriptParameter'
              - $ref: '#/components/schemas/IntegerScriptParameter'
              - $ref: '#/components/schemas/NumberArrayScriptParameter'
              - $ref: '#/components/schemas/NumberScriptParameter'
              - $ref: '#/components/schemas/StringArrayScriptParameter'
              - $ref: '#/components/schemas/StringScriptParameter'
          description: "List of arguments that will be passed to the script.\r\nIf script content defines parameters argument can be provided.\r\nThe parameter names defined in the script content should match the names specified in this list."
          nullable: true
        output_objects_format:
          $ref: '#/components/schemas/OutputObjectsFormat'
        state:
          $ref: '#/components/schemas/ScriptExecutionState'
        reason:
          type: string
          description: "Reason for the current script execution state. In most of the cases reason field will be empty. In case\r\nof an error or cancellation reason will contain information about the reason that caused script execution to\r\nbecome in this state."
          nullable: true
          readOnly: true
        start_time:
          type: string
          description: Time at which the script execution was started. String representing time in format ISO 8601.
          format: date-time
          nullable: true
          readOnly: true
        end_time:
          type: string
          description: Time at which the script execution was finished. String representing time in format ISO 8601.
          format: date-time
          nullable: true
          readOnly: true
      additionalProperties: false
      description: "Script Execution resource allows you to run a script in a runspace.\r\nThe API allows you to create, cancel, and retrieve script executions."
    BooleanArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    BooleanScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    IntegerArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    IntegerScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    NumberArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    NumberScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    StringArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    StringScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      name: X-SRS-API-KEY
      in: header
    basicAuth:
      type: http
      scheme: basic
    signAuth:
      type: http
      scheme: SIGN
security:
  - apiKeyAuth: []
  - basicAuth: []
  - signAuth: []